backend/pyproject.toml
[tool.poetry]
name = "cvm"
version = "0.1.0"
description = "Template da Python Week 22"
authors = ["Bruno Rocha <rochacbruno@gmail.com>"]
include = ["cvm/settings.toml"]

[tool.poetry.dependencies]
python = "^3.10"
sqlmodel = "^0.0.6"
fastapi = "^0.75.2"
typer = "^0.4.1"
rich = "^12.2.0"
dynaconf = "^3.1.8"
black = "^22.3.0"
flake8 = "^4.0.1"
requests = "^2.28.1"
from-root = "^1.0.2"
psycopg2-binary = "^2.9.4"
uvicorn = "^0.19.0"

[tool.poetry.dev-dependencies]
ipython = "^8.2.0"

[tool.poetry.scripts]
cvm = "cvm.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
